Checking in AdaptiveRRT for IROS 13 publication. AdaptiveRRT uses a simple cost-adaptive reward based learning scheme to select an RRT growth method. It extends OBRRT to have available to it all the appropriate growth methods in our code base. Other changes include adding an IncStat default value to 1.0. PrintMap only prints the block map if there are nodes in it now.
git-svn-id: svn+ssh://parasol-svn.cs.tamu.edu/research/parasol-svn/svnrepository/pmpl/trunk@4076 5e2c2758-810a-0410-8eb3-8e4e92d4004e

+++ src/Utilities/MetricUtils.h
@@ -101,7 +101,7 @@ class StatClass {
     // Graph Operation Statistics Accessors/Modifiers
     int GetGOStat(string _s) {return m_goStats[_s];}
     void SetGOStat(string _s, int _v) {m_goStats[_s]=_v;}
-    void IncGOStat(string _s, double _v=1) {m_goStats[_s]+=_v;}
+    void IncGOStat(string _s, int _v = 1) {m_goStats[_s]+=_v;}
 
     //Local Planner Statistics Accessors/Modifiers
     double GetLPStat(string _s){return m_lpStats[_s];}
